requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-environ==0.11.2
django-redis==5.4.0
django-solo==2.3.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.112.1
frozenlist==1.4.1
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
magic-filter==1.0.12
msgpack==1.0.8
multidict==6.0.5
oauthlib==3.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.2.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.2
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.4
zope.interface==7.0.1
